+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Changed
+ - Way popout windows opens
+
 ## 1.1.4 - 2020-06-12
 ### Added
  - Seek ability to player controls